test StreamHandler(this work with LoopEventManager):
the GUI keep crashing

teory:
the loopEventManager dispatch every and each task inside of a thread,
then almost every task is a thread of third level. the mainTread trigger the LoopEventManager(whic is a Trhead) and this trigger the task-thread.

 pyqt has pyqtSignals and pyqtSlots to handle events. I had observe that signals start to fail when they are in a third level Thread.
that is the reason why I created the loopEventManager and my own signalHandlers.

Then I think the problem could be that a taks-thread(in LoopEventManager) that update something in the GUI it  will be calling  a pyqtSlot.
And that could cause  that strange things happens.